[Fix] Move auth_utils
